app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pika
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
python-multipart
email-validator
google-genai
pillow
jinja2